@@ -6,13 +6,11 @@ import android.os.Bundle;
 
 import com.design.pattern.proxy.notification.R;
 
-public class NotifyActivity extends Activity
-{
-	
-	@Override
-	protected void onCreate(Bundle savedInstanceState)
-	{
-		super.onCreate(savedInstanceState);
-		setContentView(R.layout.activity_main);
-	}
+public class NotifyActivity extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
 }
