@@ -1,44 +1,42 @@
 package com.dm.proxy.uml.proxysubject;
 
-import com.dm.proxy.uml.activity.MainApplication;
 import com.dm.proxy.uml.subject.ILawsuit;
-import com.yline.log.LogFileUtil;
+import com.yline.utils.LogUtil;
 
-public class Lawyer implements ILawsuit
-{
+/**
+ * 静态代理的实现，需要手动继承接口；因此有直接依赖
+ *
+ * @author yline 2020-04-11 -- 12:56
+ */
+public class Lawyer implements ILawsuit {
     private ILawsuit mLawsuit;
-    
-    public Lawyer(ILawsuit lawsuit)
-    {
+
+    public Lawyer(ILawsuit lawsuit) {
         this.mLawsuit = lawsuit;
     }
-    
+
     @Override
-    public void submit()
-    {
+    public void submit() {
         mLawsuit.submit();
-        LogFileUtil.v(MainApplication.TAG, "Lawyer:" + "处理 submit");
+        LogUtil.v("Lawyer:" + "处理 submit");
     }
-    
+
     @Override
-    public void burden()
-    {
+    public void burden() {
         mLawsuit.burden();
-        LogFileUtil.v(MainApplication.TAG, "Lawyer:" + "处理 burden");
+        LogUtil.v("Lawyer:" + "处理 burden");
     }
-    
+
     @Override
-    public void defend()
-    {
+    public void defend() {
         mLawsuit.defend();
-        LogFileUtil.v(MainApplication.TAG, "Lawyer:" + "处理 defend");
+        LogUtil.v("Lawyer:" + "处理 defend");
     }
-    
+
     @Override
-    public void finish()
-    {
+    public void finish() {
         mLawsuit.finish();
-        LogFileUtil.v(MainApplication.TAG, "Lawyer:" + "处理 finish");
+        LogUtil.v("Lawyer:" + "处理 finish");
     }
-    
+
 }
